requirements.txt
streamlit
langchain-community
langchain-openai
langchain
langgraph
pymupdf
faiss-cpu
python-dotenv
isort
ruff